+++ test/foundry/testUtils/ReentrancyMocks.sol
@@ -57,7 +57,7 @@ contract ReenterBurn {
         bool reenter = !activated;
         activated = true;
         uint64 poolId;
-        uint256 vegoid = 4;
+        uint8 vegoid = 4;
 
         {
             poolId = uint40(uint160(address(this)) >> 112) + uint64(vegoid << 40);
@@ -125,7 +125,7 @@ contract ReenterMint {
         bool reenter = !activated;
         activated = true;
         uint64 poolId;
-        uint256 vegoid = 4;
+        uint8 vegoid = 4;
 
         {
             poolId = uint40(uint160(address(this)) >> 112) + uint64(vegoid << 40);
@@ -193,7 +193,7 @@ contract ReenterTransferSingle {
         bool reenter = !activated;
         activated = true;
         uint64 poolId;
-        uint256 vegoid = 4;
+        uint8 vegoid = 4;
         {
             poolId = uint40(uint160(address(this)) >> 112) + uint64(vegoid << 40);
             poolId += uint64(uint24(tickSpacing)) << 48;
@@ -260,7 +260,7 @@ contract ReenterTransferBatch {
         bool reenter = !activated;
         activated = true;
         uint64 poolId;
-        uint256 vegoid = 4;
+        uint8 vegoid = 4;
         {
             poolId = uint40(uint160(address(this)) >> 112) + uint64(vegoid << 40);
             poolId += uint64(uint24(tickSpacing)) << 48;
@@ -285,7 +285,7 @@ contract Reenter1155Initialize {
     address public token1;
     uint24 public fee;
     uint64 poolId;
-    uint256 vegoid = 4;
+    uint8 vegoid = 4;
     bool activated;
 
     function construct(address _token0, address _token1, uint24 _fee, uint64 _poolId) public {
@@ -330,7 +330,7 @@ contract Reenter1155InitializeV4 {
     uint256[65535] private __gap;
 
     PoolKey key;
-    uint256 vegoid = 4;
+    uint8 vegoid = 4;
     bool activated;
 
     function construct(PoolKey memory _key) public {
